fix(events): Custom event handler registration from other resources should add event handlers, not override existing PushEventHandlers
If a resource registers an event handler that sonoran itself needs to function properly, it will get overridden since the current registration overrides the map value. With this fix we instead add the custom event handlers to a list for each event name and calls every third-party eventhandler. This also opens up the ability for multiple third-party resources to register the same event handler without destroying registration for another resource.